fix(easyflash): 完成移植，并修复eeprom空间配置
eeprom_onRTC：修复空间错误，非256K 实际空间32k

+++ bsp/lpc408x/drivers/FM31XX/inc/fm31xx.h
@@ -46,8 +46,8 @@ extern "C"
     int E2PromProtect(uint8_t ben);
 
     /** FM31XX port */
-    int E2PromWrite(uint16_t addr, void *p, uint16_t len);
-    int E2PromRead(uint16_t addr, void *p, uint16_t len);
+    int E2PromWrite(uint16_t addr, void *p, uint32_t len);
+    int E2PromRead(uint16_t addr, void *p, uint32_t len);
 #ifdef __cplusplus
 }
 #endif
